blog_summarizer_withai/requirements.txt
openai>=1.58.0
requests>=2.32.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
streamlit>=1.40.0


